Fix `make test` to unbreak `make clean`
`make test` is known to be failing since a long time. An issue is that
it thus leaks a directory that prevents `make clean` to succeed because
of restrictive file access modes.

Separate the building of the test, which introduces the file access mode
issue, from the actual test run. This allows to fully benefit from the
access mode fixing introduced by PR #240.

Signed-off-by: Greg Kurz <groug@kaod.org>