controllers: create kata-cc-* runtime classes
This adds initial support for Confidential on baremetal. Until now,
Confidential was only supported by peerpods. With this change, we are
introducing the Confidential feature to baremetal as well.

Our consensus is to use the same config option "confidential" to enable
Confidential features in both cases.

We are also reusing the same logic for KataConfigPoolSelector and
checkNodeEligibility, just as we have been doing so far.

All the MachineConfigs will be installed by default on all nodes
alongside CRIO, regardless of whether the machine has support. This will
be RPM responsability. The RuntimeClass would be responsible for
filtering machines according to their capabilities.

Related: KATA-3967.

Signed-off-by: Beraldo Leal <bleal@redhat.com>